New Appointment

+++ app/Http/Controllers/Admin/ServiceAppointmentLayoutController.php
@@ -53,9 +53,7 @@ class ServiceAppointmentLayoutController extends AdminBaseController
         return view(parent::loadDefaultVars($this->view_path.'.create'), compact('data'));
     }
 
-
-
-   public function store(Request $request)
+    public function store(Request $request)
     {
        $data = ServiceAppointmentLayout::create([
             'title'  => $request->get('title'),
@@ -68,23 +66,58 @@ class ServiceAppointmentLayoutController extends AdminBaseController
 
     }
 
+    public function update(Request $request, $id)
+    {
+        if (!$this->idExist($id)) {
+            return redirect()->route('appointment-layout.index')->withErrors(['message' => 'Invalid Id']);
+        }
+
+        $data = $this->model;
+        $data->update([
+            'title'  => $request->get('title'),
+        ]);
+
+        $data->services()->sync($request->has('service_name')?$request->get('service_name'):[]);
+
+        AppHelper::flash('success', 'Record has been updated Successfully');
+
+
+        return redirect()->route('appointment-layout.index');
+    }
+
     public function edit($id)
     {
         $data = [];
         $data['services'] = OurService::where('status', 1)->pluck('title', 'id');
         $data['service_appointment'] = ServiceAppointmentLayout::find($id);
+        $data['service_first'] = $data['service_appointment']->services->pluck('id')->all();
 
         return view(parent::loadDefaultVars($this->view_path.'.edit'), compact('data'));
     }
 
     public function destroy($id)
     {
+        if (!$this->idExist($id)) {
+            return redirect()->route('appointment-layout.index')->withErrors(['message' => 'Invalid Id']);
+        }
 
+        $data = $this->model;
+        $data->services()->sync([]);
         $result = ServiceAppointmentLayout::destroy($id);
 
 
         AppHelper::flash('success', 'Record has deleted successfully');
-        return redirect()->route($this->scope . '.index');
+        return redirect()->route('appointment-layout.index');
+    }
+
+    /**
+     * Helper Methods
+     */
+    public function idExist($id)
+    {
+        $this->model = ServiceAppointmentLayout::find($id);
+
+        return $this->model;
     }
 
 }
